Added support for MS SQL datetimeoffset type
This feels like a waste of time... The datetimeoffset type stores
the time zone as an offset in minutes. How about storing a proper
time zone identifier instead? There is so much wrong with this, I
don't even know where to begin. But still, here it is, for the sake
of completeness, if nothing else.