Added support for generated columns
The idea is to not try to insert or update generated columns.
Unfortunartely, MS SQL doesn't reports GENERATED ALWAYS AS columns
as generated, even though they are. It does report AS columns as
generated, correctly. (Observed with MS SQL 12.00.816 and JDBC driver
11.2.0.0.) I don't see any way to work around this problem, or at
least, I don't see anything in the information schema that would help.